requirements.txt
django==2.2
django-rest-framework==3.9.2
requests==2.21.0
tornado==6.0.2
tornado_http_auth==1.1.1
